TASK: Move tests to functional and adjust
