+++ Neos.Flow/Tests/Functional/Security/Policy/PolicyTest.php
@@ -10,6 +10,7 @@ namespace Neos\Flow\Tests\Functional\Security\Policy;
  * information, please view the LICENSE file which was distributed with this
  * source code.
  */
+use Neos\Flow\Annotations as Flow;
 use Neos\Flow\Tests\FunctionalTestCase;
 
 /**
@@ -21,6 +22,13 @@ class PolicyTest extends FunctionalTestCase
      * @var boolean
      */
     protected $testableSecurityEnabled = true;
+
+
+    /**
+     * @Flow\Inject
+     * @var \Neos\Flow\Security\Policy\PolicyService
+     */
+    protected $policyService;
 
     /**
      * @test
@@ -46,5 +54,14 @@ class PolicyTest extends FunctionalTestCase
 
         self::assertTrue($this->securityContext->hasRole('Neos.Flow:Anonymous'), 'Anonymous - hasRole()');
         self::assertTrue($hasAnonymousRole, 'Anonymous - getRoles()');
+    }
+
+
+    /**
+     * @test
+     */
+    public function hasRoleReturnsTrueEvenIfNotConfigured()
+    {
+        self::assertTrue($this->policyService->hasRole('Neos.Flow:AnnotatedRole'));
     }
 }
